03_mlflow_models/requirements.txt
# MLflow Models - Python Dependencies

# Core MLflow
mlflow>=3.0

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Additional ML frameworks (notebook 03)
xgboost>=2.0.0
lightgbm>=4.0.0

# Model serving (notebook 05)
flask>=2.3.0
requests>=2.31.0

# Utilities
pyyaml>=6.0
